don't try to remove results dir unless it exists

+++ amplicon_consensus/call_bases_and_consense.py
@@ -122,8 +122,9 @@ def start():
     run(['tar', '-xvf', tarfile, '-C', raw_dir])
 
   bararr_in_results = os.path.join(results_dir, 'bararr.csv')
-  if not os.path.exists(bararr_in_results) or not filecmp.cmp(
-      bararr_in, bararr_in_results):
+  if os.path.exists(results_dir) and (
+      not os.path.exists(bararr_in_results) or not filecmp.cmp(
+        bararr_in, bararr_in_results)):
     info('bararr changed, discarding previous %s' % (
       os.path.basename(results_dir)))
     shutil.rmtree(results_dir)
